+++ buildSrc/assets/themes/danganronpa/ibuki/light/ibuki.light.hyper.definition.json
