+++ buildSrc/assets/themes/fate/rin/dark/rin.dark.hyper.definition.json
