+++ buildSrc/assets/themes/franxx/zeroTwo/dark/zero.two.dark.hyper.definition.json
