+++ buildSrc/assets/themes/franxx/zeroTwo/light/zero.two.light.hyper.definition.json
