+++ buildSrc/assets/themes/gate/rory/dark/rory.dark.hyper.definition.json
