+++ buildSrc/assets/themes/literature/monika/dark/monika.dark.hyper.definition.json
