+++ buildSrc/assets/themes/literature/yuri/dark/yuri.dark.hyper.definition.json
