+++ buildSrc/assets/themes/literature/yuri/light/yuri.light.hyper.definition.json
