+++ buildSrc/assets/themes/steinsGate/kurisu/dark/kurisu.dark.hyper.definition.json
