+++ buildSrc/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "doki-theme-hyper-build",
+  "version": "1.0.0",
+  "description": "Build process for the hyper Doki Themes",
+  "main": "index.js",
+  "repository": "git@github.com:doki-theme/doki-theme-hyper.git",
+  "author": "Alex Simons <alex@unthrottled.io>",
+  "license": "MIT",
+  "private": false,
+  "devDependencies": {
+    "@types/aws-sdk": "^2.7.0",
+    "@types/jest": "^26.0.0",
+    "@types/lodash": "^4.14.155",
+    "@types/xml2js": "^0.4.5",
+    "copy-webpack-plugin": "^6.0.2",
+    "doki-build-source": "1.0.0",
+    "jest": "^26.0.1",
+    "rimraf": "^3.0.2",
+    "ts-jest": "^26.1.0",
+    "ts-loader": "^7.0.5",
+    "ts-node": "^8.10.2",
+    "typescript": "^3.9.5",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-merge": "^4.2.2"
+  },
+  "dependencies": {
+    "aws-sdk": "^2.784.0",
+    "file-type": "^16.0.1",
+    "xml2js": "^0.4.23"
+  },
+  "scripts": {
+    "buildThemes": "ts-node ./src/BuildThemes.ts"
+  }
+}
